requirements.txt
# Polymarket Arbitrage System Dependencies

# HTTP & API
httpx>=0.25.0
aiohttp>=3.9.0
websockets>=12.0

# Blockchain / Web3
web3>=6.11.0
eth-account>=0.10.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# News & Sentiment
tweepy>=4.14.0
newsapi-python>=0.2.7
textblob>=0.17.1
vaderSentiment>=3.3.2

# Social Media Intelligence
praw>=7.7.1              # Reddit API wrapper
telethon>=1.34.0         # Telegram client

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Scheduling & Async
apscheduler>=3.10.0
asyncio-throttle>=1.0.2

# Config & Logging
python-dotenv>=1.0.0
loguru>=0.7.2
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Visualization (optional)
matplotlib>=3.8.0
plotly>=5.18.0
